requirements.txt
python-telegram-bot==13.15
huggingface_hub
Flask
Pillow
